requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.3.0
audioread==3.0.1
birdnet==0.1.7
birdnetlib==0.18.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
datasets==2.14.4
decorator==5.2.1
dill==0.3.7
evaluate==0.4.4
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
gast==0.6.0
geographiclib==2.1
geopandas==1.1.1
geopy==2.4.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
hdbscan==0.8.40
hf-xet==1.1.5
huggingface-hub==0.33.2
idna==3.10
jinja2==3.1.6
joblib==1.5.1
keras==2.15.0
keras-tuner==1.4.7
kiwisolver==1.4.8
kt-legacy==1.0.5
lazy-loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
markdown==3.8.2
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml-collections==1.1.0
ml-dtypes==0.3.2
mplcursors==0.7
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.15
networkx==3.5
numba==0.61.2
numpy==1.26.4
oauthlib==3.3.1
opencv-python-headless==4.11.0.86
opt-einsum==3.4.0
ordered-set==4.1.0
packaging==25.0
pandas==2.3.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
propcache==0.3.2
protobuf==4.25.8
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydub==0.25.1
pygments==2.19.2
pynndescent==0.5.13
pyogrio==0.11.1
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
resampy==0.4.3
rich==14.1.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.16.0
seaborn==0.13.2
segmentation-models-pytorch==0.5.0
setuptools==80.9.0
shapely==2.1.2
simsimd==6.4.9
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
stringzilla==3.12.5
sympy==1.14.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metal==1.2.0
termcolor==3.1.0
threadpoolctl==3.6.0
timm==1.0.16
tokenizers==0.21.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.0
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
watchdog==2.1.9
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.14.1
xxhash==3.5.0
yarl==1.20.1